requirements.txt
# ForgeLedger Test - Python Dependencies
# Backend framework
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0

# Data validation and settings
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Database
asyncpg>=0.29.0,<1.0.0

# Migrations
alembic>=1.13.0,<2.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
python-dateutil>=2.8.0,<3.0.0

# Testing
pytest>=7.4.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
pytest-cov>=4.1.0,<6.0.0
httpx>=0.25.0,<1.0.0

# Code quality
black>=23.0.0,<25.0.0
ruff>=0.1.0,<1.0.0
mypy>=1.7.0,<2.0.0
isort>=5.12.0,<6.0.0
